% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifwf.R
\docType{package}
\name{multifwf}
\alias{multifwf}
\alias{multifwf-package}
\title{Read Fixed Width Format Files containing lines of different Type}
\description{
Read a table of fixed width formatted data of different types into a data.frame for each type.
}
\details{
The only function you're likely to need from \pkg{multifwf} is \code{\link{read.multi.fwf}}.
}
\author{
Panos Rontogiannis \email{p.g.ronto@gmail.com}
}

